refactor: move edusharing code into own folder
